gateway/requirements.txt
fastapi==0.115.0
pydantic==2.10.2
uvicorn>=0.15.0,<0.16.0
starlette==0.38.2
python-jose[cryptography]==3.3.0
PyJWT==2.6.0
python-multipart==0.0.5
requests==2.28.1
aiofiles==0.8.0
python-dotenv==0.21.0
minio==7.1.17
pika==1.3.2
passlib[bcrypt]==1.7.4
tenacity==8.2.3
pydantic-settings==2.0.3
httpx==0.25.2
